refactor(auth): migrate from better-auth to new auth module and implement requireAuthentication utility
