+++ web-app/src/app/(app)/components/sidebar/components/agents-list.tsx
@@ -1,6 +1,4 @@
 import { AgentListType } from "@prisma/client";
-import { headers } from "next/headers";
-import { redirect } from "next/navigation";
 import { getTranslations } from "next-intl/server";
 import { Suspense } from "react";
 
@@ -14,7 +12,7 @@ import {
   SidebarMenuItem,
 } from "@/components/ui/sidebar";
 import { Skeleton } from "@/components/ui/skeleton";
-import { auth } from "@/lib/better-auth/auth";
+import { requireAuthentication } from "@/lib/auth/utils";
 import { getOrCreateAgentListsByTypes } from "@/lib/db/services/agentList.service";
 
 function AgentsListSkeleton() {
@@ -45,16 +43,9 @@ function AgentsListSkeleton() {
 
 async function AgentsListContent() {
   const t = await getTranslations("App.Sidebar.Content.AgentsList");
-  const session = await auth.api.getSession({
-    headers: await headers(),
-  });
+  const { session } = await requireAuthentication();
 
-  const userId = session?.user.id;
-  if (!userId) {
-    redirect("/signin");
-  }
-
-  const agentLists = await getOrCreateAgentListsByTypes(userId, [
+  const agentLists = await getOrCreateAgentListsByTypes(session.user.id, [
     AgentListType.FAVORITE,
   ]);
