+++ web-app/src/app/(landing)/(auth)/reset-password/actions.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { auth } from "@/lib/better-auth/auth";
+import { auth } from "@/lib/auth/auth";
 
 import { resetPasswordFormSchema, ResetPasswordFormSchemaType } from "./data";
