+++ web-app/src/lib/actions/auth.actions.ts
@@ -1,5 +1,5 @@
 "use server";
-import { auth } from "@/lib/better-auth/auth";
+import { auth } from "@/lib/auth/auth";
 
 export async function signInSocial(
   provider: "google" | "microsoft" | "apple" | "linkedin",
