+++ web-app/src/lib/auth/auth.client.ts
